+++ cli/tests/unit/test_auth.py
@@ -14,23 +14,15 @@ def test_no_browser_auth():
         assert op.call_count == 0
 
 
-def test_when_authenticated():
-    with patch('dcos.auth.force_auth'):
-
-        _mock_dcos_run([util.which('dcos')], True)
-        assert auth.force_auth.call_count == 0
-
-
 def test_anonymous_login():
     with patch('sys.stdin.readline', return_value='\n'), \
             patch('uuid.uuid1', return_value='anonymous@email'):
 
-        assert _mock_dcos_run([util.which('dcos'),
-                               'help'], False) == 0
-        assert _mock_dcos_run([util.which('dcos'), 'config',
-                               'show', 'core.email'], False) == 0
-        assert _mock_dcos_run([util.which('dcos'), 'config',
-                               'unset', 'core.email'], False) == 0
+        assert _mock_dcos_run(['dcos', 'help'], False) == 0
+        assert _mock_dcos_run(['dcos',
+                               'config', 'show', 'core.email'], False) == 0
+        assert _mock_dcos_run(['dcos',
+                               'config', 'unset', 'core.email'], False) == 0
 
 
 def _mock_dcos_run(args, authenticated=True):
@@ -41,6 +33,13 @@ def _mock_dcos_run(args, authenticated=True):
 
     with patch('sys.argv', args), patch.dict(os.environ, env):
         return main()
+
+
+def test_when_authenticated():
+    with patch('dcos.auth.force_auth'):
+
+        _mock_dcos_run(['dcos'], True)
+        assert auth.force_auth.call_count == 0
 
 
 def _config_with_credentials():
