+++ cli/tests/unit/test_http_auth.py
