+++ cli/tox.ini
@@ -1,17 +1,17 @@
 [tox]
-envlist = syntax, py{27,34}-unit, py{27,34}-integration
+envlist = py34-syntax, py{27,34}-unit, py{27,34}-integration
 
 [testenv]
 deps =
   teamcity-messages
+  mock
   pytest
   pytest-cov
-  mock
   pytz
   -e..
 passenv = DCOS_* CI_FLAGS CLI_TEST_SSH_KEY_PATH CLI_TEST_MASTER_PROXY
 
-[testenv:syntax]
+[testenv:py34-syntax]
 deps =
   teamcity-messages
   flake8
