+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = syntax, py{27,34}-unit
+envlist = py34-syntax, py{27,34}-unit
 
 [testenv]
 deps =
@@ -8,7 +8,7 @@ deps =
   pytest-cov
 passenv = DCOS_* CI_FLAGS CLI_TEST_SSH_KEY_PATH
 
-[testenv:syntax]
+[testenv:py34-syntax]
 deps =
   teamcity-messages
   flake8
